+++ js/data.js
@@ -1,0 +1,8 @@
+const DATA_TYPES =['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const DATA_TIMES =['12:00', '13:00', '14:00'];
+const DATA_FEATURES =['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const DATA_PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+export {DATA_TYPES, DATA_TIMES, DATA_FEATURES, DATA_PHOTOS};
